Delete generate spec fix; Rewrite generate spec intention
